feat: enable fast restart from sleep
Adds a module to store json objects in sleep memory. We take a task dict, convert it to a json
string, then convert that to a bytearray. Before sleeping, the contents of the task bytearray are
written to sleep memory. On startup, we not check the sleep memory to see if there is a saved task.
If there is, we immediately display it while we continue the wifi connect and task refresh steps.
This means you almost immediately get a task displayed on wake. Regular boots are unaffected as the
sleep memory contains no useful data.